packages/frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Pretend frontend project"
readme = "README.md"
authors = [
    { name = "dhaustein" }
]
requires-python = ">=3.13"
dependencies = []

[project.optional-dependencies]
test = [
    "playwright>=1.55.0",
    "pytest-playwright>=0.7.1",
]

[project.scripts]
frontend = "frontend:main"

[build-system]
requires = ["uv_build>=0.8.8,<0.9.0"]
build-backend = "uv_build"
